tech(api): migrate particpant result controller/route to its bounded-context
