+++ api/src/prescription/campaign-participation/infrastructure/serializers/jsonapi/participant-result-serializer.js
