+++ api/tests/prescription/campaign-participation/integration/application/campaign-participation-route_test.js
@@ -1,11 +1,15 @@
 import { campaignParticipationController } from '../../../../../src/prescription/campaign-participation/application/campaign-participation-controller.js';
 import * as moduleUnderTest from '../../../../../src/prescription/campaign-participation/application/campaign-participation-route.js';
+import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Integration | Application | Route | campaignParticipationRouter', function () {
-  let httpTestServer;
+  let httpTestServer, sandbox;
 
   beforeEach(async function () {
+    sandbox = sinon.createSandbox();
+    sandbox.stub(securityPreHandlers, 'checkRequestedUserIsAuthenticatedUser');
+
     sinon.stub(campaignParticipationController, 'getAnalysis').callsFake((request, h) => h.response('ok').code(200));
     sinon
       .stub(campaignParticipationController, 'getCampaignAssessmentParticipation')
@@ -13,9 +17,46 @@ describe('Integration | Application | Route | campaignParticipationRouter', func
     sinon
       .stub(campaignParticipationController, 'getCampaignAssessmentParticipationResult')
       .callsFake((request, h) => h.response('ok').code(200));
+    sinon
+      .stub(campaignParticipationController, 'getUserCampaignAssessmentResult')
+      .callsFake((request, h) => h.response('ok').code(200));
 
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
+  });
+
+  afterEach(function () {
+    sandbox.restore();
+  });
+
+  describe('#getUserCampaignAssessmentResult', function () {
+    context('Error cases', function () {
+      it('should not called controller when user not authenticated', async function () {
+        // given
+        securityPreHandlers.checkRequestedUserIsAuthenticatedUser.callsFake((request, h) => {
+          return Promise.resolve(h.response().code(403).takeover());
+        });
+
+        // when
+        httpTestServer.request('GET', '/api/users/1234/campaigns/5678/assessment-result');
+
+        // then
+        expect(campaignParticipationController.getUserCampaignAssessmentResult.notCalled).to.be.true;
+      });
+
+      it('should return a 401 HTTP response', async function () {
+        // given
+        securityPreHandlers.checkRequestedUserIsAuthenticatedUser.callsFake((request, h) => {
+          return Promise.resolve(h.response().code(401).takeover());
+        });
+
+        // when
+        const response = await httpTestServer.request('GET', '/api/users/1234/campaigns/5678/assessment-result');
+
+        // then
+        expect(response.statusCode).to.equal(401);
+      });
+    });
   });
 
   describe('GET /api/campaign-participations/{id}/analyses', function () {
